When creating a new account, the system tries to find if any existing user may already exist, and warns the user against creating a twin for him/herself.
Since the algorithm may fail, one can enforce the user creation
by clicking again on submit.

Links to the "twins" can point to users that are not subscribed to the
current conference. The system should therefore allow linking to users
not registered to the current conference (it doesn't allow this at
this time).


git-svn-id: svn://svn.mongueurs.net/act/trunk@873 67b57a05-4208-db11-a765-00306e02d86a