Add a "sidetrack" to the schedule
For ye2010, they defined parts of the partner/extra program as 'out'
events, but they occured in parallel with some of the talks (master
classes).

This calls for the creation of a new "room": sidetrack, for events
that happen in parallel with talks or events happening in normal rooms,
but should be left unchanged by so-called "global events" happening in
'venue' or 'out'.

The translation of the sidetrack room name in done in the template,
because it wasn't possible to load Act::I18N in Act::Config (circular
dependencies).

This will allow to put the "partners program" in the schedule for
future conference.

Of course, this fixes the issue for all conferences, but ye2010,
because they overrode templates/core/talk/schedule with their
own version. :-(



git-svn-id: svn://svn.mongueurs.net/act/trunk@2231 67b57a05-4208-db11-a765-00306e02d86a