Link the PAUSE id to MetaCPAN
MetaCPAN has a more complete profile page where CPAN users can link to
all their other accounts (Github, StackOverflow...), so this is more
useful as a CPAN author page than search.cpan.org.
Also link the "PAUSE id" label to the PAUSE site as this may not be
obvious to Perl beginners that come to the conferences.