Added updatemytalks_as private handler and sub ajax_stateful_handler
Added 'updatemytalks_as' for ajax_stateful_handler to specifically set or clean user's
favorite task, not to toggle.

On some pages of YAPC::EU::2015 there is few 'favorite' buttons for same task,
and when they both off and visitor presses them one-by-one, they both become
toggled to 'on' state, but two in a row request to old updatemytalks_a handler
toggles task in favorites from off to on and back to off.

So we need separate handler:
* updatemytalks_a -> toggles by 'task_id' in request.
* updatemytalks_as -> sets 'task_id' favorite state depending from additional 'state' parameter

For latter one we already provide preparations in javascript already, there:
https://github.com/Act-Conferences/ye2015/blob/master/wwwdocs/js/main.js#L22-L26