Make password case sensitive, except if it contains only lower cases
Up to that commit, passwords are lc-ed before being hashed and stored.
In order to authentify users with "old" passwords, if the submitted
password is first tested as is, and if the test fails, it is lc-ed
then tested.

Hence as long as the password stored in the DB contains only lower
case, we cannot know if it's because its an old password or if it's
a new one containing only lower cases.

And we cannot change the case of the password at login if the test
works after lc-ed the password because maybe the "correct" password
is all lower case.